pyproject.toml
[tool.poetry]
name = "petsdrf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg = {extras = ["binary"], version = "^3.1.12"}
djangorestframework = "^3.14.0"
django = "^4.2.7"
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.2"
pylint-django = "^2.5.5"
mypy = "^1.6.1"
django-stubs = "^4.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
